Cast standard notation literal to `uint32_t`
Technically, this is a `double` -> `uint32_t` narrowing conversion, but
since I was invoking the conversion with `5e8` (in range of a 32-bit
unsigned integer), I wouldn't have lost information.

Regardless, good practice to spell out data conversions.

+++ rosbag2_transport/test/rosbag2_transport/test_play_timing.cpp
@@ -194,7 +194,7 @@ TEST_F(PlayerTestFixture, play_ignores_invalid_delay)
 
 TEST_F(PlayerTestFixture, play_respects_starting_time)
 {
-  rclcpp::Duration start_delay(0, 5e8);  // 5e8 ns == 0.5 s
+  rclcpp::Duration start_delay(0, static_cast<uint32_t>(5e8));  // 5e8 ns == 0.5 s
   rclcpp::Duration delay_margin(1, 0);
 
   // Start 5 seconds into the bag
